Nuke old interface from orbit.
I killed the old smac interface, the submit_call stuff
and the cli interface (which still lives on in deprecated because I was not sure
whether some of you need it). We have to review how the metalearning process
works with these changes as I am not sure how it works in general :).