Drop lockfile requirement
The lockfile requirement is no longer necessary since the drop of
pSMAC for parallel SMAC/Auto-sklearn as all parallelism is now
controlled by a central Auto-sklearn instance which saves the
datamanager and the ensemble targets to disk.

As a positive side-effect this removes a potential deadlock in the
ensemble builder.