Readded desired code for design choice on model indexing
There are two indexes that can be used, SMAC uses `config_id` and asklearn
uses `num_run`, these are not garunteed to be equal and also `num_run`
is not always present.

As the user should not care that there is possible 2 indexes for models,
made the choice to show `config_id` as this allows displaying info on
failed runs.

An alternative to show asklearn's `num_run` index is just to exclude any
failed runs from showing up in the leaderboard.