Fix the ensemble_size == 0 error in automl.py (#1369)
* Fix the ensemble == 0 error in fit_ensemble and show_models function by adding a valueError to the former and giving a warning and returning empty dictionary in the latter

* Update automl.py

* Two tests for ensemble_size == 0 cases

Added two tests to check if the automl.fit_ensemble() raises error when ensemble_size == 0 and if show_models() returns empty dictionary when ensemble_size == 0

* Update automl.py

* Update test_automl.py

Test for checking if the show_models() functions raise an error if models are not fitted.

* Update automl.py

Add a function __sklearn_is_fitted__() which returns the boolean value of self.fitted(). And add the check for model fitting in show_models() function.

* Update autosklearn/automl.py

* Formatting changes to clear all the pre-commit tests

Co-authored-by: Eddie Bergman <eddiebergmanhs@gmail.com>