Fix: `leaderboard` fills nans (#1432)
For models that are in the ensemble but have no corresponding entry in
model_runs (built from run_history), we fill in NA values to prevent
crashing. Why this occurs, I don't know but this should at least prevent
total failure when we can still provide other useful information back to
the user.